Added constant file.
